+++ examples/ConnectedInteractiveChart.tsx
@@ -87,6 +87,7 @@ const CHART = (
     }}
     className='example-chart'
     loadDataDebounceTimeout={0}
+    pixelRatio={window.devicePixelRatio || 1}
   >
     <Stack>
       <ConnectedSimpleLineLayer seriesId={SERIES_ID}/>