+++ src/connected/export-only/exportableActions.ts
@@ -1,3 +1,5 @@
+import { ActionType } from '../flux/atomicActions';
+
 export {
   setYDomains,
   setHover,
@@ -7,3 +9,7 @@ export {
 export {
   setXDomainAndLoad as setXDomain
 } from '../flux/compoundActions';
+
+export function getActionTypeName(actionValue: any): string {
+  return ActionType[actionValue] || null;
+}